+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added Link List Block.
 - Added Hero style for Media & Text Block.
 - Added more icons and changed icons to come from HDS.
+- Added Search Block.
 - Updated Helsinki Design System version.
 - Updated navigation styles.
 - Updated footer styles and content.